removed old FastRPTest class
